Update README.md
Adding and reordering to make sure users are going to build their project only to discover they have to do the TMP fix for something to work. Unlikely really, but potentially possible. Intent is to include a "Save the project now as a template" kind of situation so the user isn't repeating this whole guide.